third_party/WebKit/LayoutTests/platform/linux/virtual/outofblink-cors-http/http/tests/fetch/serviceworker/thorough/redirect-credentials-expected.txt

This is a testharness.js-based test.
PASS Startup 
PASS redirect-credentials 
PASS executeTest-0 
PASS executeTest-1 
PASS executeTest-2 
PASS executeTest-3 
PASS executeTest-4 
PASS executeTest-5 
FAIL executeTest-6 assert_equals: http://127.0.0.1:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F%26ACAOrigin%3Dhttp%3A%2F%2F127.0.0.1%3A8000&mode=cors&credentials=omit&method=GET should be resolved expected "resolved" but got "rejected"
FAIL executeTest-7 assert_equals: http://127.0.0.1:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F%26ACAOrigin%3Dhttp%3A%2F%2F127.0.0.1%3A8000%26ACACredentials%3Dtrue&mode=cors&credentials=omit&method=GET should be resolved expected "resolved" but got "rejected"
PASS executeTest-8 
FAIL executeTest-9 assert_equals: http://127.0.0.1:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F%26ACAOrigin%3Dhttp%3A%2F%2F127.0.0.1%3A8000%26ACACredentials%3Dtrue&mode=cors&credentials=include&method=GET should be resolved expected "resolved" but got "rejected"
FAIL executeTest-10 assert_equals: http://127.0.0.1:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F%26ACAOrigin%3Dhttp%3A%2F%2F127.0.0.1%3A8000&mode=cors&credentials=same-origin&method=GET should be resolved expected "resolved" but got "rejected"
FAIL executeTest-11 assert_equals: http://127.0.0.1:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2Flocalhost%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3F%26ACAOrigin%3Dhttp%3A%2F%2F127.0.0.1%3A8000%26ACACredentials%3Dtrue&mode=cors&credentials=same-origin&method=GET should be resolved expected "resolved" but got "rejected"
FAIL executeTest-12 assert_equals: http://localhost:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2F127.0.0.1%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*&mode=cors&credentials=omit&method=GET&ACAOrigin=* should be resolved expected "resolved" but got "rejected"
PASS executeTest-13 
FAIL executeTest-14 assert_equals: http://localhost:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2F127.0.0.1%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3D*&mode=cors&credentials=same-origin&method=GET&ACAOrigin=* should be resolved expected "resolved" but got "rejected"
FAIL executeTest-15 assert_equals: http://localhost:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2F127.0.0.1%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3Dnull%26ACACredentials%3Dtrue&mode=cors&credentials=omit&method=GET&ACAOrigin=http://127.0.0.1:8000&ACACredentials=true should be resolved expected "resolved" but got "rejected"
FAIL executeTest-16 assert_equals: http://localhost:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2F127.0.0.1%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3Dnull%26ACACredentials%3Dtrue&mode=cors&credentials=include&method=GET&ACAOrigin=http://127.0.0.1:8000&ACACredentials=true should be resolved expected "resolved" but got "rejected"
FAIL executeTest-17 assert_equals: http://localhost:8000/serviceworker/resources/redirect.php?Redirect=http%3A%2F%2F127.0.0.1%3A8000%2Fserviceworker%2Fresources%2Ffetch-access-control.php%3FACAOrigin%3Dnull%26ACACredentials%3Dtrue&mode=cors&credentials=same-origin&method=GET&ACAOrigin=http://127.0.0.1:8000&ACACredentials=true should be resolved expected "resolved" but got "rejected"
Harness: the test ran to completion.
